Remove ConfigurationOption class and use ConfigOption enum instead.
