+++ alitheia/core/src/main/resources/hibernate.cfg.xml
@@ -50,7 +50,6 @@
         <mapping class="eu.sqooss.service.db.Branch"/>
         <mapping class="eu.sqooss.service.db.StoredProject"/>
         <mapping class="eu.sqooss.service.db.StoredProjectConfig"/>
-        <mapping class="eu.sqooss.service.db.ConfigurationOption"/>
         <mapping class="eu.sqooss.service.db.ClusterNode"/>
         <mapping class="eu.sqooss.service.db.Plugin"/>
         <mapping class="eu.sqooss.service.db.PluginConfiguration"/>
